pyproject.toml
[tool.poetry]
name = "megano"
version = "0.1.0"
description = ""
authors = ["Vitaliy Zhilyaev <jilyaev1987@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
django-debug-toolbar = "^4.2.0"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
